feat: 优化丁香园肺炎路由 (#3828)
