+++ docs/other.md
@@ -402,17 +402,9 @@ type 为 all 时，category 参数不支持 cost 和 free
 
 <Route author="DIYgod" example="/coronavirus/dxy" path="/coronavirus/dxy"/>
 
-### 丁香园 - 全国新型肺炎疫情数据统计（全国）
+### 丁香园 - 全国新型肺炎疫情数据统计
 
-<Route author="DIYgod" example="/coronavirus/dxy/data" path="/coronavirus/dxy/data"/>
-
-### 丁香园 - 全国新型肺炎疫情数据统计（各省份）
-
-<Route author="DIYgod" example="/coronavirus/dxy/data/湖北" path="/coronavirus/dxy/data/:province" :paramsDesc="['省份名']"/>
-
-### South China Morning Post - China coronavirus outbreak
-
-<Route author="DIYgod" example="/coronavirus/scmp" path="/coronavirus/scmp"/>
+<Route author="DIYgod HenryQW" example="/coronavirus/dxy/data/湖北" path="/coronavirus/dxy/data/:province?" :paramsDesc="['省份，缺省为全国']"/>
 
 ## 新趣集
