+++ lib/router.js
@@ -2155,8 +2155,7 @@ router.get('/weixin/miniprogram/release', require('./routes/weixinMiniprogram/re
 
 // 武汉肺炎疫情动态
 router.get('/coronavirus/caixin', require('./routes/coronavirus/caixin'));
-router.get('/coronavirus/dxy/data/:province', require('./routes/coronavirus/dxy-data-province'));
-router.get('/coronavirus/dxy/data', require('./routes/coronavirus/dxy-data-country'));
+router.get('/coronavirus/dxy/data/:province?', require('./routes/coronavirus/dxy-data'));
 router.get('/coronavirus/dxy', require('./routes/coronavirus/dxy'));
 router.get('/coronavirus/scmp', require('./routes/coronavirus/scmp'));
 router.get('/coronavirus/nhc', require('./routes/coronavirus/nhc'));
